#123 - Repository interface extracted
